% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_markers.R
\name{calculate_markers}
\alias{calculate_markers}
\title{calculate_markers}
\usage{
calculate_markers(
  corpus,
  file,
  id = "id",
  text = "text",
  show_progress = interactive()
)
}
\arguments{
\item{corpus}{A \code{\link[base]{data.frame}} containing the text to process.}

\item{file}{The marker file. All marker files are external to the library for licensing purposes. See the vignettes for help.}

\item{id}{The name of the PK column for the \code{corpus}.}

\item{text}{The names of text column in \code{corpus} to process.}

\item{show_progress}{Should a progress bar be shown?}
}
\value{
the annotated corpus
}
\description{
Calculate the linguistic markers for a corpus
}
\author{
Mark Newman, \email{mark@trinetteandmark.com}
}
